+++ modules/flok/src/notistack-lib/utils.tsx
@@ -25,3 +25,12 @@ export function apiNotification(
     },
   }
 }
+export function replaceDashes(str: string) {
+  let strArray = str.split("")
+  strArray.forEach((char, i) => {
+    if (char === "-") {
+      strArray[i] = " "
+    }
+  })
+  return strArray.join("")
+}
